Moved get_html to parser.py
