+++ app/parser.py
@@ -1,20 +1,29 @@
+import urllib2
+
+import util
+
+
+def get_html(url):
+    """
+    Retrieves html text from a string-formatted url
+    :param url: web url string
+    :return: html text
+    """
+    if '://' not in url:
+        formatted_url = 'http://' + url
+    else:
+        formatted_url = url
+    try:
+        return urllib2.urlopen(formatted_url).read()
+    except urllib2.URLError:
+        util.warning('Invalid URL request')
+        return None
+
+
 def parse_html(html):
     """
     Takes HTML and does some basic parsing
     :param html: html text
     :return: ???
     """
-    pass
-
-
-def url_to_dictionary(url):
-    """
-    FOR TESTING PURPOSES ONLY, takes a url and runs the various parsing functions to return a dictionary in the JSON
-    representation that Miriam's autograder accepts
-    :param url: url of requested recipe
-    :return: dictionary in autograder-acceptable format
-    """
-    from app.app import get_html
-    html = get_html(url)
-    # etc
     pass
